Enable HTTP serve mode
Add an HTTP serve mode in which `poule` can be used without any NSQ
infrastructure. In this mode, the tool listens on the specified address
for incoming GitHub webhooks.
