+++ .gitignore
@@ -1,3 +1,4 @@
 bin/
 pkg/
 /poule
+.DS_Store
